compiles, but im not sure this optimization can be done before column lifetime analysis. parent projection maps are invalidated when children flip
