+++ src/optimizer/join_order/join_order_optimizer.cpp
@@ -10,18 +10,6 @@
 
 namespace duckdb {
 
-static bool HasJoin(LogicalOperator *op) {
-	while (!op->children.empty()) {
-		if (op->children.size() == 1) {
-			op = op->children[0].get();
-		}
-		if (op->children.size() == 2) {
-			return true;
-		}
-	}
-	return false;
-}
-
 unique_ptr<LogicalOperator> JoinOrderOptimizer::Optimize(unique_ptr<LogicalOperator> plan,
                                                          optional_ptr<RelationStats> stats) {
 
@@ -59,13 +47,6 @@ unique_ptr<LogicalOperator> JoinOrderOptimizer::Optimize(unique_ptr<LogicalOpera
 			new_logical_plan->estimated_cardinality = relation_stats.at(0).cardinality;
 			new_logical_plan->has_estimated_cardinality = true;
 		}
-	}
-
-	// only perform left right optimizations when stats is null (means we have the top level optimize call)
-	// Don't check reorderability because non-reorderable joins will result in 1 relation, but we can
-	// still switch the children.
-	if (stats == nullptr && HasJoin(new_logical_plan.get())) {
-		new_logical_plan = query_graph_manager.LeftRightOptimizations(std::move(new_logical_plan));
 	}
 
 	// Propagate up a stats object from the top of the new_logical_plan if stats exist.
