+++ src/optimizer/join_order/query_graph_manager.cpp
@@ -335,106 +335,4 @@ unique_ptr<LogicalOperator> QueryGraphManager::RewritePlan(unique_ptr<LogicalOpe
 	return plan;
 }
 
-static void FlipChildren(LogicalOperator &op) {
-	std::swap(op.children[0], op.children[1]);
-	if (op.type == LogicalOperatorType::LOGICAL_COMPARISON_JOIN || op.type == LogicalOperatorType::LOGICAL_DELIM_JOIN) {
-		auto &join = op.Cast<LogicalComparisonJoin>();
-		join.join_type = InverseJoinType(join.join_type);
-		for (auto &cond : join.conditions) {
-			std::swap(cond.left, cond.right);
-			cond.comparison = FlipComparisonExpression(cond.comparison);
-		}
-	}
-	if (op.type == LogicalOperatorType::LOGICAL_ANY_JOIN) {
-		auto &join = op.Cast<LogicalAnyJoin>();
-		join.join_type = InverseJoinType(join.join_type);
-	}
-}
-
-void QueryGraphManager::TryFlipChildren(LogicalOperator &op, idx_t cardinality_ratio) {
-	auto &left_child = op.children[0];
-	auto &right_child = op.children[1];
-	auto lhs_cardinality = left_child->has_estimated_cardinality ? left_child->estimated_cardinality
-	                                                             : left_child->EstimateCardinality(context);
-	auto rhs_cardinality = right_child->has_estimated_cardinality ? right_child->estimated_cardinality
-	                                                              : right_child->EstimateCardinality(context);
-	if (rhs_cardinality < lhs_cardinality * cardinality_ratio) {
-		return;
-	}
-	FlipChildren(op);
-}
-
-unique_ptr<LogicalOperator> QueryGraphManager::LeftRightOptimizations(unique_ptr<LogicalOperator> input_op) {
-	auto op = input_op.get();
-	// pass through single child operators
-	while (!op->children.empty()) {
-		if (op->children.size() == 2) {
-			switch (op->type) {
-			case LogicalOperatorType::LOGICAL_COMPARISON_JOIN: {
-				auto &join = op->Cast<LogicalComparisonJoin>();
-
-				switch (join.join_type) {
-				case JoinType::INNER:
-				case JoinType::OUTER:
-					TryFlipChildren(join);
-					break;
-				case JoinType::LEFT:
-				case JoinType::RIGHT:
-					if (join.right_projection_map.empty()) {
-						TryFlipChildren(join, 2);
-					}
-					break;
-				case JoinType::SEMI:
-				case JoinType::ANTI: {
-					idx_t has_range = 0;
-					if (!PhysicalPlanGenerator::HasEquality(join.conditions, has_range)) {
-						// if the conditions have no equality, do not flip the children.
-						// There is no physical join operator (yet) that can do a right_semi/anti join.
-						break;
-					}
-					TryFlipChildren(join, 2);
-					break;
-				}
-				default:
-					break;
-				}
-				break;
-			}
-			case LogicalOperatorType::LOGICAL_CROSS_PRODUCT: {
-				// cross product not a comparison join so JoinType::INNER will get ignored
-				TryFlipChildren(*op, 1);
-				break;
-			}
-			case LogicalOperatorType::LOGICAL_ANY_JOIN: {
-				auto &join = op->Cast<LogicalAnyJoin>();
-				if (join.join_type == JoinType::LEFT && join.right_projection_map.empty()) {
-					TryFlipChildren(join, 2);
-				} else if (join.join_type == JoinType::INNER) {
-					TryFlipChildren(join, 1);
-				}
-				break;
-			}
-			case LogicalOperatorType::LOGICAL_DELIM_JOIN: {
-				auto &join = op->Cast<LogicalComparisonJoin>();
-				if (HasInverseJoinType(join.join_type) && join.right_projection_map.empty()) {
-					FlipChildren(join);
-					join.delim_flipped = true;
-				}
-				break;
-			}
-			default:
-				break;
-			}
-			op->children[0] = LeftRightOptimizations(std::move(op->children[0]));
-			op->children[1] = LeftRightOptimizations(std::move(op->children[1]));
-			// break from while loop
-			break;
-		}
-		if (op->children.size() == 1) {
-			op = op->children[0].get();
-		}
-	}
-	return input_op;
-}
-
 } // namespace duckdb
