+++ src/optimizer/optimizer.cpp
@@ -4,6 +4,7 @@
 #include "duckdb/main/client_context.hpp"
 #include "duckdb/main/config.hpp"
 #include "duckdb/main/query_profiler.hpp"
+#include "duckdb/optimizer/build_probe_side_optimizer.hpp"
 #include "duckdb/optimizer/column_lifetime_optimizer.hpp"
 #include "duckdb/optimizer/common_aggregate_optimizer.hpp"
 #include "duckdb/optimizer/compressed_materialization.hpp"
@@ -176,9 +177,10 @@ unique_ptr<LogicalOperator> Optimizer::Optimize(unique_ptr<LogicalOperator> plan
 		column_lifetime.VisitOperator(*plan);
 	});
 
-	// swaps left and right children of joins so that the table with less data is the build side
+	// Once we know the column lifetime, we have more information regarding
+	// what relations should be the build side/probe side.
 	RunOptimizer(OptimizerType::BUILD_SIDE_PROBE_SIDE, [&]() {
-		BuildProbeSideOptimizer build_probe_side_optimizer(true);
+		BuildProbeSideOptimizer build_probe_side_optimizer(context);
 		build_probe_side_optimizer.VisitOperator(*plan);
 	});
